Support loading the OUI db from a compiled-in asset.
This allows users to build binaries with the database embedded within,
so not requiring it be loaded from the filesystem separately. Note
that if the database changes, the following command should be run

$ go generate

This will rebuild assets.go, which contains the OUI db. The go-bindata
package is required for this:

$ go get -u github.com/jteeuwen/go-bindata/...